configure script: no longer uses regex operator
Some versions of bash seem to lack support for the regex (=~) operator and
always return false rather than erroring out. This causes the template files
to be erroneously included in the dependencies. Using a string comparison
with wildcards should accomplish the same thing on a wider variety of bash
versions: detect 'template' in the filename and exclude it.